unbound: fix local_subnet for IPv6 addresses that contain a port
This prevents a forwarding server named like ::1@5453 from being added
to unbound.conf as a forward-host instead of the correct forward-addr.
forward-host requires the name to be resolved, which is impossible in
the absence of another nameserver. Thus, forwarding-only configurations
referencing only the IPv6 loopback address with a port number were
broken.

Signed-off-by: Mark Mentovai <mark@mentovai.com>